+++ src/Resources/contao/config/config.php
@@ -22,19 +22,6 @@ $GLOBALS['TL_HOOKS']['loadDataContainer']['multiColumnEditor']  = ['huh.multicol
 $GLOBALS['MULTI_COLUMN_EDITOR']['rsce_fields'] = [];
 
 /**
- * Assets
- */
-if (\Contao\System::getContainer()->get('huh.utils.container')->isBackend()) {
-    $GLOBALS['TL_CSS']['contao-multi-column-editor-bundle'] = 'bundles/heimrichhannotcontaomulticolumneditor/contao-multi-column-editor-bundle-be.css|static';
-}
-else
-{
-    $GLOBALS['TL_JAVASCRIPT']['sortablejs'] = 'assets/sortablejs/sortablejs/Sortable.min.js|static';
-}
-
-$GLOBALS['TL_JAVASCRIPT']['contao-multi-column-editor-bundle'] = 'bundles/heimrichhannotcontaomulticolumneditor/contao-multi-column-editor-bundle.js|static';
-
-/**
  * Ajax
  */
 $GLOBALS['AJAX'][\HeimrichHannot\MultiColumnEditorBundle\Widget\MultiColumnEditor::NAME] = [